+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add `LocalSpans::to_span_records()`.
+- Add `#[trace(properties = { "k1": "v1", "k2": "v2" })]`.
 
 ## v0.6.2